# record: ensure block.written is accessed atomically