@nicktrav 2022-05-23
This is just one approach to solving this issue - honoring the contract that the struct field comment implies re: atomicity.

It looks like there are also some comments around using the commit pipeline mutex for external synchronization ([for example](https://github.com/cockroachdb/pebble/blob/bb2c1501ac23ca3ce17d46717986610424b05790/record/log_writer.go#L582)), though evidently something is slipping through the cracks are we're seeing the race in Cockroach.

For context, in cockroachdb/cockroach#81133, parallel writes are made to each engine's WAL as part of liveness checking, which can race if the context is cancelled and the log is closed.

Seems like this code in Pebble has been this way since inception, and we're only tickling the issue now due to the nature of the concurrent access introduced in the Cockroach PR.