@petermattis 2022-05-24
`LogWriter.{Write,SyncRecord,Close}` need external synchronization. Inside Pebble this synchronization is provided by `commitPipeline.mu`. It looks like we're not holding `commitPipeline.mu` correctly in the `DB.Close()` path.